Ollama_Approach/requirements.txt
streamlit
python-dotenv
PyPDF2
pdfplumber
langchain